NXP-30742: clean up enrichers to not pollute logs with warns

+++ modules/platform/nuxeo-platform-tag/src/main/java/org/nuxeo/ecm/platform/tag/io/TagsJsonEnricher.java
@@ -49,15 +49,17 @@ public class TagsJsonEnricher extends AbstractJsonEnricher<DocumentModel> {
 
     @Override
     public void write(JsonGenerator jg, DocumentModel document) throws IOException {
-        TagService tagService = Framework.getService(TagService.class);
-        jg.writeArrayFieldStart(NAME);
         try (SessionWrapper wrapper = ctx.getSession(document)) {
+            if (!wrapper.getSession().exists(document.getRef())) {
+                return;
+            }
+            TagService tagService = Framework.getService(TagService.class);
+            jg.writeArrayFieldStart(NAME);
             for (String tag : tagService.getTags(wrapper.getSession(), document.getId())) {
                 jg.writeString(tag);
             }
+            jg.writeEndArray();
         }
-        jg.writeEndArray();
-
     }
 
 }
